remove obsolete code and update reporting scripts
