+++ cmd/kubectl-frisbee/commands/common/constants.go
@@ -16,6 +16,12 @@ limitations under the License.
 
 package common
 
+import (
+	"time"
+
+	"k8s.io/apimachinery/pkg/util/wait"
+)
+
 const (
 	ClusterScope        = ""
 	FrisbeeInstallation = "frisbee"
@@ -26,3 +32,18 @@ const (
 	FrisbeeRepo  = "https://carv-ics-forth.github.io/frisbee/charts"
 	JetstackRepo = "https://charts.jetstack.io"
 )
+
+const (
+	ManagedNamespace = "app.kubernetes.io/managed-by=Frisbee"
+)
+
+const (
+	TestTimeout = "24h"
+)
+
+var BackoffPodCreation = wait.Backoff{
+	Duration: 10 * time.Second,
+	Factor:   1.5,
+	Jitter:   0.1,
+	Steps:    60,
+}
