+++ examples/federated_learning/node-placement/install.sh
@@ -2,6 +2,7 @@
 
 set -eu
 set -o pipefail
+trap "trap - SIGTERM && kill -- -$$" SIGINT SIGTERM EXIT
 
 export NAMESPACE=node-placement
 export SCENARIO=$(dirname -- "$0")/manifest.yml
@@ -11,13 +12,13 @@ export DEPENDENCIES=(./charts/system/ ./charts/federated-learning/fedbed/)
 # Prepare the Reporting folder
 mkdir -p "${REPORTS}"
 
-# Submit the scenario and follow server logs
-kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}" --logs server |& tee -a "${REPORTS}"/logs &
-
 # Copy the manifest
 cp "${SCENARIO}" "${REPORTS}"
 
-# wait for the scenario to be submitted
+# Submit the scenario and follow server logs
+kubectl-frisbee submit test "${NAMESPACE}" "${SCENARIO}" "${DEPENDENCIES[@]}" --logs server |& tee  "${REPORTS}"/logs &
+
+# Give a headstart
 sleep 10
 
 # Report the scenario
