--- pkg/client/interface.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2022-2023 ICS-FORTH.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package client
-
-/*
-// Client is the Frisbee API client abstraction
-type Client interface {
-	TestManagementAPI
-	TestInspectionAPI
-}
-
-// TestManagementAPI describes scenario api methods
-type TestManagementAPI interface {
-	// GetScenario queries the Kubernetes API for the given Scenario id.
-	GetScenario(testName string) (scenario *v1alpha1.Scenario, err error)
-
-	// ListScenarios queries the Kubernetes API the selected labels.
-	ListScenarios(selector string) (tests v1alpha1.ScenarioList, err error)
-
-	// ListVirtualObjects  list all virtual objects with the given selector
-	ListVirtualObjects(namespace string, selectors ...string) (list v1alpha1.VirtualObjectList, err error)
-
-	// DeleteTest deletes single test by name
-	DeleteTest(testName string) error
-
-	// DeleteTests deletes all tests with the selected labels.
-	DeleteTests(selector string) (testNames []string, err error)
-}
-
-type TestInspectionAPI interface {
-	// Logs returns logs stream from job pods, based on job pods logs
-	// Logs(testName string) (logs chan output.Output, err error)
-}
-
-
-*/
